download files atomically
We now download to temporary files before rename()ing them into
their proper places. This reduces the effects of partial downloads due
to (e.g.) network issues. We also add a new key to the memo,
latest_complete_master, which is the last master version we know we
have a full set of databases for.